backend/planarian/src/planarian/pyproject.toml
[tool.poetry]
name = "planaria-backend"
version = "0.1.0"
description = "AI Agent Builder Backend using CrewAI"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
crewai = "^0.80.0"
crewai-tools = "^0.12.0"
openai = "^1.52.0"
anthropic = "^0.39.0"
google-generativeai = "^0.8.0"
python-dotenv = "^1.0.0"
pydantic = "^2.9.0"
jinja2 = "^3.1.4"
fastapi = "^0.115.0"
uvicorn = "^0.32.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"